+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.1] - 2026-03-04
+
+### Added
+
+- Folder path autocomplete in settings for `Markdown template folder`
+
+### Changed
+
+- `Markdown template folder` now supports typeahead selection inspired by Templater-style folder picking
+
 ## [1.5.0] - 2026-03-04
 
 ### Added